refactor(servicebus): move NimBusReceiverHostedService to NimBus.ServiceBus.Hosting
Promotes the Service-Bus-typed receiver out of NimBus.SDK and into the
ServiceBus transport project as ServiceBusReceiverHostedService +
ServiceBusReceiverOptions, with a new AddServiceBusReceiver DI extension.

The SDK keeps NimBusReceiverHostedService / NimBusReceiverOptions / the
AddNimBusReceiver extension as [Obsolete] bridges that subclass / forward
to the new types so existing consumers keep compiling. AddNimBusReceiver
also TryAddSingletons IServiceBusAdapter -> ISubscriberClient so the new
receiver resolves the adapter without a type-name dependency on the SDK.

Slice 1 of #18 (drop Azure.Messaging.ServiceBus from NimBus.SDK). Cheapest
move per the abstractions-builder handoff — standalone, easy revert. The
SDK still references NimBus.ServiceBus and Azure.Messaging.ServiceBus;
those drop in subsequent slices.

ResolverWorker sample migrated to AddServiceBusReceiver so the
warning-as-error gate doesn't trip on the sample's [Obsolete] consumption.
