+++ samples/AspirePubSub/AspirePubSub.ResolverWorker/Program.cs
@@ -1,6 +1,5 @@
 using NimBus.Resolver;
-using NimBus.SDK.Hosting;
-using NimBus.ServiceBus;
+using NimBus.ServiceBus.Hosting;
 
 var builder = Host.CreateApplicationBuilder(args);
 
@@ -8,20 +7,11 @@ builder.AddServiceDefaults();
 builder.Configuration["ResolverId"] = "Resolver";
 builder.Services.AddResolver();
 
-builder.Services.AddHostedService(sp =>
+builder.Services.AddServiceBusReceiver(options =>
 {
-    var client = sp.GetRequiredService<Azure.Messaging.ServiceBus.ServiceBusClient>();
-    var adapter = sp.GetRequiredService<IServiceBusAdapter>();
-    var logger = sp.GetRequiredService<ILogger<NimBusReceiverHostedService>>();
-
-    var options = new NimBusReceiverOptions
-    {
-        TopicName = "Resolver",
-        SubscriptionName = "Resolver",
-        MaxConcurrentSessions = 8
-    };
-
-    return new NimBusReceiverHostedService(client, adapter, options, logger);
+    options.TopicName = "Resolver";
+    options.SubscriptionName = "Resolver";
+    options.MaxConcurrentSessions = 8;
 });
 
 var host = builder.Build();
